feat: [ENG-1786] add the ConsumeQueue to the poisonQueueHandler to allow consuming the poison queue (#68)
Add the logic to consume the poison queue
This logic makes sure to take all the result up to the latest offset when the consumption is done.
The consumption record count is non-deterministic because we are consuming by batch the records and we complete when the new batch includes the final offset define when the consumption started. We can end up with a small amount of records on top of the final offset, but the logic makes sure that we stop when we include that offset.
The logic intend to make a single execution with all the records to the handler. This is to prevent scenario where we can push back more event on the queue and accidentally consume them.