feat: [ENG-1726] Add the configuration and logic to use manual commit instead of auto-commit (#79)
This is the first iteration of improvement for the manual commit logic for the consumer. The current implementation is trivial as it execute all the records the consumer received in the poll batch and then only commit all the record that were processed at the end. This allows to prevent issue such as :

Network failure while processing that prevent the record to actually be handled
Node failure where the instance that was executing the logic simply fails and never reach completion
Current usage of the consumer is mostly only using a single partition, therefore the rebalancing while polling issue is not currently an issue. But to ensure better performance and reduce the possibility of processing multiple time a same event, we should look into a followup to allow async execution of the batch and to handle partition rebalancing.