chore: [ENG-3245] add different utils to reduce boilerplate (#112)
This is a smashup of multiple utility to reduce code duplication and increase readability.

### Must

Reduces boilerplate in tests for code that never returns error and is used to generate expected data

### logger.WithSpanStartOptions

Help instrumentation to match our spans in our code

### OutputErrsMatchInputLength

Unify the way we handle batch method outputs to identify if there is an error or not.

### Instrument

Boilerplate reducer for instrumentation, also allows to use the pattern in a function execution and not only on a struct method
 
### ComponentName

Ensure the same format is used for component names